+++ packages/web/CHANGELOG.md
@@ -1,5 +1,9 @@
 # @floating-ui-plus/web
 
+## 0.7.2
+
+## 0.7.1
+
 ## 0.7.0
 
 ### Minor Changes